Finished exercises for week2

+++ week2/model.php
@@ -1,9 +1,9 @@
 <?php
 /**
  * Model
- * User: reinardvandalen
- * Date: 05-11-18
- * Time: 15:25
+ * User: Folkert Leistra
+ * Date: 28-11-2018
+ * Time: 17:30
  */
 
 /* Enable error reporting */
@@ -97,7 +97,7 @@ function get_breadcrumbs($breadcrumbs) {
  * @param array $navigation Array with as Key the page name and as Value the corresponding url
  * @return string html code that represents the navigation
  */
-function get_navigation($navigation){
+function get_navigation($template, $active_id){
     $navigation_exp = '
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <a class="navbar-brand">Series Overview</a>
@@ -106,15 +106,16 @@ function get_navigation($navigation){
     </button>
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav mr-auto">';
-    foreach ($navigation as $name => $info) {
-        if ($info[1]){
+    foreach ($template as $name => $info) {
+        /* makes the active url bold */
+        if ($name == $active_id) {
             $navigation_exp .= '<li class="nav-item active">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
-        }else{
-            $navigation_exp .= '<li class="nav-item">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="' . $template[$active_id]['url'] . '">' . $template[$active_id]['name'] . '</a>';
         }
-
+        else {
+            $navigation_exp .= '<li class="nav-item">';
+            $navigation_exp .= '<a class="nav-link" href="' . $info['url'] . '">' . $info['name'] . '</a>';
+        }
         $navigation_exp .= '</li>';
     }
     $navigation_exp .= '
@@ -126,25 +127,31 @@ function get_navigation($navigation){
 
 /**
  * Creats a Bootstrap table with a list of series
- * @param object $db pdo object
+ * @param string $user name of the user
  * @param array $series with series from the db
  * @return string
  */
-function get_serie_table($series){
+function get_serie_table($series, $db){
+
     $table_exp = '
     <table class="table table-hover">
     <thead
     <tr>
         <th scope="col">Series</th>
-        <th scope="col"></th>
+        <th scope="col">Added by</th>
+        <th scope="col">Information</th>
     </tr>
     </thead>
     <tbody>';
     foreach($series as $key => $value){
+
         $table_exp .= '
         <tr>
             <th scope="row">'.$value['name'].'</th>
+            <td>'.get_name($db,$value['user']).'</td>
             <td><a href="/DDWT18/week2/serie/?serie_id='.$value['id'].'" role="button" class="btn btn-primary">More info</a></td>
+            
+            
         </tr>
         ';
     }
@@ -211,6 +218,7 @@ function get_serieinfo($pdo, $serie_id){
  * @return string
  */
 function get_error($feedback){
+    $feedback = json_decode($feedback, True);
     $error_exp = '
         <div class="alert alert-'.$feedback['type'].'" role="alert">
             '.$feedback['message'].'
@@ -224,7 +232,8 @@ function get_error($feedback){
  * @param array $serie_info post array
  * @return array with message feedback
  */
-function add_serie($pdo, $serie_info){
+function add_serie($pdo, $serie_info) {
+    $user_id = get_user_id();
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -258,12 +267,13 @@ function add_serie($pdo, $serie_info){
     }
 
     /* Add Serie */
-    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract) VALUES (?, ?, ?, ?)");
+    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract, user) VALUES (?, ?, ?, ?, ?)");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
-        $serie_info['Abstract']
+        $serie_info['Abstract'],
+        $user_id
     ]);
     $inserted = $stmt->rowCount();
     if ($inserted ==  1) {
@@ -286,7 +296,16 @@ function add_serie($pdo, $serie_info){
  * @param array $serie_info post array
  * @return array
  */
-function update_serie($pdo, $serie_info){
+function update_series($pdo, $serie_info){
+    $stmt = $pdo->prepare('SELECT user FROM series WHERE id = ?');
+    $stmt->execute([$serie_info['serie_id']]);
+    $id = $stmt->fetch();
+    if ($id['user'] != get_user_id()) {
+        return [
+            'type' => 'danger',
+            'message' => 'You do not have the privilege to edit that'
+        ];
+    }
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -298,6 +317,7 @@ function update_serie($pdo, $serie_info){
         return [
             'type' => 'danger',
             'message' => 'There was an error. Not all fields were filled in.'
+
         ];
     }
 
@@ -333,7 +353,8 @@ function update_serie($pdo, $serie_info){
         $serie_info['Creator'],
         $serie_info['Seasons'],
         $serie_info['Abstract'],
-        $serie_info['serie_id']
+        $serie_info['serie_id'],
+
     ]);
     $updated = $stmt->rowCount();
     if ($updated ==  1) {
@@ -356,13 +377,18 @@ function update_serie($pdo, $serie_info){
  * @param int $serie_id id of the to be deleted series
  * @return array
  */
-function remove_serie($pdo, $serie_id){
-    /* Get series info */
-    $serie_info = get_serieinfo($pdo, $serie_id);
-
+function remove_serie($pdo, $serie_info){
+    $user_id = get_user_id();
+    if (($user_id != $serie_info['user'])){
+        return [
+            'type' => 'danger',
+            'message' => 'You do not have the privileges to delete that serie.'
+        ];
+    }
     /* Delete Serie */
     $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
-    $stmt->execute([$serie_id]);
+    $stmt->execute([$serie_info['id']]);
+
     $deleted = $stmt->rowCount();
     if ($deleted ==  1) {
         return [
@@ -392,6 +418,19 @@ function count_series($pdo){
 }
 
 /**
+ * Count the number of series listed on Series Overview
+ * @param object $pdo database object
+ * @return mixed
+ */
+function count_users($pdo){
+    /* Get series */
+    $stmt = $pdo->prepare('SELECT * FROM users');
+    $stmt->execute();
+    $series = $stmt->rowCount();
+    return $series;
+}
+
+/**
  * Changes the HTTP Header to a given location
  * @param string $location location to be redirected to
  */
@@ -405,10 +444,189 @@ function redirect($location){
  * @return bool current user id or False if not logged in
  */
 function get_user_id(){
-    session_start();
+    if (session_status() == PHP_SESSION_NONE) {
+        session_start();
+    }
     if (isset($_SESSION['user_id'])){
         return $_SESSION['user_id'];
     } else {
         return False;
     }
+}
+
+/**
+ * Get name of user by ID
+ * @param object $pdo database object
+ * @parim $id int
+ * @return array
+ */
+
+function get_name($pdo, $id){
+    $stmt = $pdo->prepare('SELECT firstname, lastname FROM users WHERE id = ?');
+    $stmt->execute([$id]);
+    $result = $stmt->fetch();
+    $name = $result['firstname'] .' '. $result['lastname'];
+    return $name;
+
+}
+
+/**
+ * Register a new user to the database
+ * @param object $pdo database object
+ * @parim array $form_data with POST data
+ * @return array
+ */
+function register_user($pdo, $form_data){
+
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password']) or
+        empty($form_data['firstname']) or
+        empty($form_data['lastname'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username, password, first- and last name.'
+        ];
+    }
+    /* Check if user already exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_exists = $stmt->rowCount();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for existing username */
+    if ( !empty($user_exists) ) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does already exists!'
+        ];
+    }
+    /* Hash password */
+    $password = password_hash($form_data['password'], PASSWORD_DEFAULT);
+    /* Save user to the database */
+    try {
+        $stmt = $pdo->prepare('INSERT INTO users (username, password, firstname, lastname) VALUES (?, ?, ?, ?)');
+        $stmt->execute([$form_data['username'], $password, $form_data['firstname'],
+            $form_data['lastname']]);
+        $user_id = $pdo->lastInsertId();
+    } catch (PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Login user and redirect */
+    session_start();
+    print_r($user_id);
+    $_SESSION['user_id'] = $user_id;
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf('%s, your account was successfully created!', get_name($pdo, $_SESSION['user_id']))
+    ];
+    redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+        json_encode($feedback)));
+}
+
+/**
+ * login a user to the database
+ * @param object $pdo database object
+ * @parim array $form_data with POST data
+ * @return array
+ */
+function login_user($pdo, $form_data){
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username and password.'
+        ];
+    }
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_info = $stmt->fetch();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for wrong username */
+    if ( empty($user_info) ) {
+        return  [
+            'type' => 'danger',
+            'message' => 'The username you entered does not exist!'
+        ];
+    }
+    /* Check password */
+    if ( !password_verify($form_data['password'], $user_info['password']) ){
+        return [
+            'type' => 'danger',
+            'message' => 'The password you entered is incorrect!'
+        ];
+    } else {
+        session_start();
+        $_SESSION['user_id'] = $user_info['id'];
+        $feedback = [
+            'type' => 'success',
+            'message' => sprintf('%s, you were logged in successfully!',
+                get_name($pdo, $_SESSION['user_id']))
+        ];
+        redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+            json_encode($feedback)));
+
+    }
+}
+
+/**
+ * Checks if a user is currently logged in
+ * @return bool
+ */
+function check_login(){
+    if (session_status() == PHP_SESSION_NONE) {
+        session_start();
+    }
+    if (isset($_SESSION['user_id'])){
+        return True;
+    } else {
+        return False;
+    }
+}
+
+/**
+ * logs out an user by destroying the session
+ *
+ */
+function logout_user(){
+    if (session_status() == PHP_SESSION_NONE) {
+        session_start();
+    }
+    session_destroy();
+    if (!check_login()) {
+        $feedback = [
+            'type' => 'success',
+            'message' => 'You were logged out successfully!'
+        ];
+        redirect(sprintf('/DDWT18/week2/?error_msg=%s',
+            json_encode($feedback)));
+    }
+    else {
+        $feedback = [
+            'type' => 'fail',
+            'message' => 'There was a problem logging you out.'
+        ];
+        redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+            json_encode($feedback)));
+    }
+
 }
